modificar conexiones para que sean async

+++ repositories/user_repository.py
@@ -1,14 +1,18 @@
 from models.user import *
 from sanic import response
-from config.database import db
+from config.database import get_database
 
-user_collection = db["users"]
-
-def create_user(user):
+async def create_user(user):
     validate = user.validate()
     
     if validate:
         return response.json({'error': validate}, status=400)
+    
+    db = await get_database()        
+    if db is None:
+        raise Exception("Fallo al establecer conexion con la BD")
+    user_collection = db["users"]
+    
     existing_user = user_collection.find_one({"email": user.email})
     if existing_user:
         return response.json({'error':"El correo electrónico ya está en uso"}, status=400)
@@ -19,19 +23,37 @@ def create_user(user):
     return User.from_dict(user_dict)
 
 
-def read_users():
+async def read_users():
+    db = await get_database()        
+    if db is None:
+        raise Exception("Fallo al establecer conexion con la BD")
+    user_collection = db["users"]
+    
+    user_collection = db["users"]
     users = user_collection.find()
     return [User.from_dict(user) for user in users]
 
 
-def read_user(user_id):
+async def read_user(user_id):
+    db = await get_database()        
+    if db is None:
+        raise Exception("Fallo al establecer conexion con la BD")
+    user_collection = db["users"]
+
+    user_collection = db["users"]
     user = user_collection.find_one({'_id': user_id})
     if user:
         return User.from_dict(user)
     else:
         return None
     
-def read_user_email(email):
+async def read_user_email(email):
+    db = await get_database()        
+    if db is None:
+        raise Exception("Fallo al establecer conexion con la BD")
+    user_collection = db["users"]
+
+    user_collection = db["users"]
     user = user_collection.find_one({'email': str(email)})
     if user:
         return User.from_dict(user)
@@ -39,9 +61,16 @@ def read_user_email(email):
         return None
 
 
-def update_user(user_id, user):
+async def update_user(user_id, user):
     user_dict = user.to_dict()
     dict_partial_update = {"$set": {key: value for key, value in user_dict.items() if value is not None and key in {'email', 'password', 'name'}}}
+    
+    db = await get_database()        
+    if db is None:
+        raise Exception("Fallo al establecer conexion con la BD")
+    user_collection = db["users"]
+
+    user_collection = db["users"]
     result = user_collection.update_one({'_id': user_id}, dict_partial_update)
     if result.modified_count == 1:
         updated_user = user_collection.find_one({'_id': user_id})
@@ -50,7 +79,15 @@ def update_user(user_id, user):
         return None
 
 
-def delete_user(user_id):
+async def delete_user(user_id):
+
+    db = await get_database()        
+    if db is None:
+        raise Exception("Fallo al establecer conexion con la BD")
+    user_collection = db["users"]
+
+    user_collection = db["users"]
+
     result = user_collection.delete_one({'_id': user_id})
     if result.deleted_count == 1:
         return True
